requirements.txt
python-dotenv
numpy>=1.26.0
pandas==2.0.3
requests